buy additional stock
buy additional stock

+++ Assignment12v2/src/stock/StockService.java
@@ -8,7 +8,7 @@ package stock;
 public interface StockService {
 	public double getPrice(Stock stock);
 	
-	public long getVolume(Stock stock);
+	public int getVolume(Stock stock);
 	
 	public double getOpenPrice(Stock stock);
